line_agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repobuddy-line-agent"
version = "0.1.0"
description = "RepoBuddy voice agent for Cartesia Line platform"
dependencies = [
    "cartesia-line>=0.1.0",
    "httpx>=0.24.0",
    "typing-extensions>=4.0.0",
]

[tool.setuptools]
py-modules = ["main", "tools"]
